botCode/requirements.txt
pyTelegramBotAPI
mysql-connector-python
pickledb
boto3
PyYAML
Flask
gunicorn